+++ src/impls.rs
@@ -1,1 +1,2 @@
+mod add;
 mod index;
